fix(holder): c_rank 改 SMALLINT，兼容 STYLE=07 超过 127 的名次

+++ tables/tb_fd_holder_structure/insert_temp.py
@@ -34,7 +34,7 @@
     CREATE TABLE tytdata.tb_fd_holder_top10 (
         c_report_date  DATE          NOT NULL,
         c_fd_code      VARCHAR(20)   NOT NULL,
-        c_rank         TINYINT       NOT NULL,
+        c_rank         SMALLINT      NOT NULL,
         c_style        VARCHAR(20),
         c_holder_name  VARCHAR(300),
         c_holder_type  VARCHAR(30),
